PCB Layout 90% Done
All that is remaining is the LA traces (moved them and forgot to re-do them) and the SPI and SHDN lines for the digital pots and regs.
Also, do not merge this branch if you can. Have it fully replace the previous folder. There were obsolete files that I removed that Github keeps because it's dumb.