Finished PCB but...
Somehow managed to break the schematic file. I know how to fix it, but it is tedious and pointless for now (one has to go through every single individual component and verify that it is the same in the PCB, since the annotation script fucked everything up in the schematic, and therefore any new Netlist).