MAJOR CHANGE! READ BELOW
Due to layout issues, I changed the pins used for the variable regulator shutdown lines. +V1_SHDN is now AA9, +V2_SHDN is now AA8, -V1_SHDN is now AA11, and -V2_SHDN is now AA10. MAKE SURE TO CHANGE THE RESPECTIVE CODE (Paolo).

Included in this upload is the pre-design review finish PCB and Schematic, with pdf files of each for printing and easy reference.